+++ release_notes.md
@@ -1,6 +1,7 @@
 ## New Features
 
 ## Bug fixes
+- Skip constructor of AzureStorageDurabilityProvider if not used, to avoid spurious validation exceptions
 
 ## Breaking Changes